requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0
enum34==1.1.6
Flask==0.12.2
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
paramiko==2.2.1

psycopg2==2.7.5
python-dotenv==0.10.3

py2app==0.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.1.2
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.22
Werkzeug==0.12.2
gunicorn==19.7.1
Pyrebase==3.0.27
firebase-admin==2.8.0
google-api-core==0.1.4
google-auth==1.4.0
google-auth-oauthlib==0.2.0
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-cloud-storage==1.7.0
sendgrid==6.0.5
cloudinary==1.14.0
